refactor: Extra test mixins to separate file
